4.4.2112 [sx_sdk_4_4_2100, febfdf79e209efa1846d7450bd2e980f180b99f2]

+++ include/linux/mlx_sx/auto_registers/mlxsw_item.h
@@ -77,7 +77,7 @@ __mlxsw_item_offset(const struct mlxsw_item *item, unsigned short index,
         BUG();
     }
 
-    return ((item->offset + item->step * index + item->in_step_offset) /
+    return ((item->offset + (unsigned int)item->step * index + item->in_step_offset) /
         typesize);
 }
